scripts/requirements.txt
beautifulsoup4==4.9.3
google-cloud-firestore==2.1.1
praw==7.2.0
requests==2.25.1
python-dotenv==0.17.1